Reverted code from save locations feature to working. Tests for RealTImeEventSearchAPI passed
